Revert file rename + finish lowerCamel const cleanup
