+++ server/src/external/tinybird/sendEvents/sendEvents.ts
@@ -3,7 +3,7 @@ import type { EventInsert } from "@autumn/shared";
 import * as Sentry from "@sentry/bun";
 import type { Logger } from "@/external/logtail/logtailUtils.js";
 import { tinybirdIngest } from "../initTinybird.js";
-import { tinybirdSecondaryApi } from "../initTinybirdSecondary.js";
+import { tinybirdSecondaryApi } from "../initTinybirdV2.js";
 import { mapToTinybirdEvent } from "./mapEvent.js";
 
 /** Generate a unique error ID for tracking */
